requirements.txt
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==1.0.1
numpy==1.23.4
opencv-python==4.6.0.66
Pillow==9.2.0
tensorflow==2.10.0
